% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCENS.LLOQ.R
\name{addCENS.LLOQ}
\alias{addCENS.LLOQ}
\title{Add a CENS column according to LLOQ.}
\usage{
addCENS.LLOQ(.data, LLOQ = 0.1, rm = TRUE)
}
\arguments{
\item{.data}{Data table formatted for Monolix WITHOUT CENS column.}

\item{LLOQ}{Lower Limit Of Quantification.}

\item{rm}{Remove previous CENS column. Default to TRUE.}
}
\description{
Add a CENS column according to LLOQ.
}
\examples{
# addCENS.LLOQ(TAB, LLOQ=.01)
}
\keyword{censoring}
