% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCENS.ULOQ.R
\name{addCENS.ULOQ}
\alias{addCENS.ULOQ}
\title{Add a CENS column according to ULOQ.}
\usage{
addCENS.ULOQ(.data, ULOQ = 100, rm = TRUE)
}
\arguments{
\item{.data}{Data table formatted for Monolix WITHOUT CENS column.}

\item{ULOQ}{Value of the upper limit of quantification.}

\item{rm}{Remove previous CENS column. Default to TRUE.}
}
\description{
Add a CENS column according to ULOQ.
}
\examples{
# addCENS.ULOQ(TAB, ULOQ = 300)
}
\keyword{censoring}
