% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.timeC.R
\name{forest.timeC}
\alias{forest.timeC}
\title{A function forest plot the Time spent vs. survival.}
\usage{
forest.timeC(project.file = "./monolix/181011_2cCLk0.mlxtran",
  c.name = "C1", survtab = PFS, simtimes = 7, c.range = c(5, 10,
  20, 30, 40, 50, 75, 100, 150, 200))
}
\arguments{
\item{project.file}{Monolix (2018) project file.}

\item{c.name}{Name of the concentration in the mlxtran project.}

\item{survtab}{Table of survival: 3 columns ID, TIME, PROG.}

\item{simtimes}{Time of evaluation}

\item{c.range}{Concentrations}
}
\description{
This function allows you to plot...
}
\keyword{monolix}
