% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx.IWRES.R
\name{mlx.IWRES}
\alias{mlx.IWRES}
\title{This function plots a publication-ready IWRES distribution of a monolix project run.}
\usage{
mlx.IWRES(
  project.dir = "../monolix/",
  project.name = "",
  drug = NULL,
  y = NULL,
  type = "pdf",
  Color = "black",
  size = 0.3
)
}
\arguments{
\item{project.dir}{Absolute or relative name of the folder of the monolix project.}

\item{project.name}{Name of the monolix project file without *.mlxtran* extension. The project must be in *project.dir*}

\item{drug}{Text for the legend. Typicaly name of the drug. Can be a **TeX** object from package **latex2exp**}

\item{y}{Name of the Obs values.}

\item{type}{Does the IWRES plot is *"pdf"* or *"cdf"*.}

\item{Color}{Color of the bars or dots. Default to *black*.}

\item{size}{Size of the dots (for type = "cdf"). Default to *0.3*.}
}
\description{
This function plots a publication-ready IWRES distribution of a monolix project run.
}
\examples{
# getwd()
}
\keyword{monolix}
