% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx.OvsP.R
\name{mlx.OvsP}
\alias{mlx.OvsP}
\title{This function plots a publication-ready OBS vs. PRED of a monolix project run.}
\arguments{
\item{project.dir}{Absolute or relative name of the folder of the monolix project.}

\item{project.name}{Name of the monolix project file without *.mlxtran* extension. The project must be in *project.dir*}

\item{drug}{Text for the legend. Typicaly name of the drug. Can be a **TeX** object from package **latex2exp**}

\item{y}{Name of the Obs values.}

\item{type}{Does the Obs vs Pop plot is *pop* or *ind*.}

\item{disp}{Display in *lin* or *log*. Default to *lin*.}

\item{Color}{Color of the dots. Default to *black*.}

\item{size}{Size of the dots. Default to *0.3*.}
}
\description{
This function plots a publication-ready OBS vs. PRED of a monolix project run.
}
\examples{
# getwd()
}
\keyword{monolix}
