% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx.VPC.R
\name{mlx.VPC}
\alias{mlx.VPC}
\title{A function to plot VPCs from a Monolix project.}
\usage{
mlx.VPC(
  project.dir = "../monolix/",
  project.name = "",
  drug = NULL,
  y = NULL,
  disp = "lin",
  obs = TRUE,
  size = 0.5,
  time.unit = "day",
  simulated = FALSE,
  corrected = FALSE,
  colors = c("#4682B4", "#FF0000", "#0080FF")
)
}
\arguments{
\item{project.dir}{Absolute or relative name of the folder of the monolix project.}

\item{project.name}{Name of the monolix project file without *.mlxtran* extension.. The project must be in *project.dir*}

\item{drug}{Text for the legend. Typicaly name of the drug. Can be a **TeX** object from package **latex2exp**}

\item{y}{Name of the Obs values.}

\item{disp}{Display in *lin* or *log*. Default to *lin*.}

\item{obs}{Plot observed data?. Default to *TRUE*.}

\item{size}{Size of the observed data dots. Default to *0.5*. Useless if *obs = FALSE*.}

\item{time.unit}{Time unit for the x legend. Default to *"day"*.}

\item{simulated}{Use simulated Blq data? Default to *FALSE*.}

\item{corrected}{Dose the predictions are corrected? Default to *FALSE*.}

\item{colors}{Color palette : order: c(lines and dots, central ribbon, lower and uper ribbon). Default to *c("#4682B4", "#FF0000", "#0080FF")*.}
}
\description{
This function allows you to have almost a publication ready VPC plot of a monolix run.
}
\examples{
# getwd()
}
\keyword{monolix}
