% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlxind.params.R
\name{mlxind.params}
\alias{mlxind.params}
\title{Get individual params (EBEs) from a Monolix run.}
\usage{
mlxind.params(project.dir = "../monolix/", project.name = "", covs = NA)
}
\arguments{
\item{project.dir}{Absolute or relative name of the folder of the monolix project.}

\item{project.name}{Name of the monolix project file without *.mlxtran* extension.. The project must be in *project.dir*.}

\item{covs}{Optional covariates to include in the extraxtion (vector of strings).}
}
\description{
Get individual params (EBEs) from a Monolix run.
}
\keyword{monolix}
