% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlxpop.print.kable.R
\name{mlxpop.print.kable}
\alias{mlxpop.print.kable}
\title{A function to export Populationparameters.txt file of a Monolix project to xlsx file.}
\usage{
mlxpop.print.kable(
  mlxproj,
  estim = "sa",
  output = "xlsx",
  caption = "Pop params",
  greek = F
)
}
\arguments{
\item{mlxproj}{Data Table.}

\item{estim}{"sa" or "lin". default to "sa".}

\item{output}{Logical. output .xlsx file or not. default to TRUE.}

\item{caption}{Caption of the table.}

\item{greek}{Convert to greek letters.}
}
\description{
This function allows you to have almost a publication ready table of the population parameter of a model project.
}
\examples{
# getwd()
}
\keyword{monolix}
