% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlxpop.table.R
\name{mlxpop.table}
\alias{mlxpop.table}
\title{A function to to print Populationparameters.txt file of a Monolix project to Rmarksdown.}
\usage{
mlxpop.table(
  project.dir = "../monolix/",
  project.name = "",
  estim = "sa",
  caption = "Population parameters of the PK model.",
  greek = FALSE
)
}
\arguments{
\item{project.dir}{Absolute or relative name of the folder of the monolix project.}

\item{project.name}{Name of the monolix project file without *.mlxtran* extension.. The project must be in *project.dir*}

\item{estim}{"sa" or "lin". default to "sa".}

\item{caption}{Caption of the table.}

\item{greek}{Convert to greek letters, "latex_engine: xelatex" is needed. default to FALSE.}
}
\description{
This function allows you to have almost a publication ready table of the population parameter of a model project.
}
\examples{
# getwd()
}
\keyword{monolix}
