% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlxpop.xlsx.R
\name{mlxpop.xlsx}
\alias{mlxpop.xlsx}
\title{A function to export Populationparameters.txt file of a Monolix project to xlsx file.}
\usage{
mlxpop.xlsx(mlxproj, estim = "sa", caption = "Pop params", greek = FALSE)
}
\arguments{
\item{mlxproj}{Name of the monolix project. The project must be in ./monolix/}

\item{estim}{"sa" or "lin". default to "sa".}

\item{caption}{Caption of the table.}

\item{greek}{Convert to greek letters, default to FALSE.}
}
\description{
This function allows you to have almost a publication ready table of the population parameter of a model project.
}
\examples{
# getwd()
}
\keyword{monolix}
