% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MlxPop2Xlsx.R
\name{MlxPop2Xlsx}
\alias{MlxPop2Xlsx}
\title{A function to export Populationparameters.txt file of a Monolix project to xlsx file.}
\usage{
MlxPop2Xlsx(mlxproj, output = TRUE)
}
\arguments{
\item{mlxproj}{Data Table.}

\item{output}{Logical. output .xlsx file or not. default to TRUE.}
}
\description{
This function allows you to have almost a publication ready table of the population parameter of a model project.
}
\keyword{monolix}
