% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotValid.R
\name{PlotValid}
\alias{PlotValid}
\title{Plot raw data for validation.}
\usage{
PlotValid(id = NULL, data = NULL, x.name = "Time (day)", y.name = NULL)
}
\arguments{
\item{id}{Patient's data to plot. First if not mentioned}

\item{data}{PK table.}

\item{x.name}{Name of the X axis.}

\item{y.name}{Name of the Y axis. e.g. "[Drug] (mg/L)"}
}
\description{
Plot raw data for validation.
}
\examples{
# PlotValid(id="1001",data=PK)
}
\keyword{data}
\keyword{plot}
\keyword{raw}
