% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Valid.Plot.R
\name{Valid.Plot}
\alias{Valid.Plot}
\title{A function to plot raw data in pdf before modelling}
\usage{
Valid.Plot(TAB, ADM, FileName, ADM2 = NA)
}
\arguments{
\item{TAB}{Data Table.}

\item{ADM}{Which ADM to plot.}

\item{FileName}{Output pdf file name.}

\item{ADM2}{Second ADM to plot on right axis (optional).}
}
\description{
This function allows you to have a first visualisation of data file.
}
\examples{
Valid.Plot()
}
\keyword{monolix}

