% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx.OvsP.R
\name{mlx.OvsP}
\alias{mlx.OvsP}
\title{A function to export Populationparameters.txt file of a Monolix project to xlsx file.}
\usage{
mlx.OvsP(mlxproj, drug = "Y", pop = FALSE)
}
\arguments{
\item{mlxproj}{Name of the monolix project. The project must be in ./monolix/}

\item{drug}{Text for the legend.}

\item{pop}{Logical. does the Obs vs Pop have to be plotted.}
}
\description{
This function allows you to have almost a publication ready table of the population parameter of a model project.
}
\examples{
# getwd()
}
\keyword{monolix}
