% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx.VPC.R
\name{mlx.VPC}
\alias{mlx.VPC}
\title{A function to plot VPCs from a Monolix project.}
\usage{
mlx.VPC(
  project.dir = "/home/nicolas/Documents/Etudes/Thymo/monolix/",
  project.name = "190916_190722_THYMO_VALID_retro",
  drug = NULL,
  y = NULL,
  disp = "lin",
  size = 0.5,
  simulated = FALSE,
  corrected = FALSE
)
}
\arguments{
\item{project.dir}{Absolute name of the folder of the monolix project.}

\item{project.name}{Name of the monolix project. The project must be in *project.dir*}

\item{drug}{Text for the legend. Typicaly name of the drug. Can be a **TeX** object from package **latex2exp**}

\item{y}{Name of the Obs values.}

\item{disp}{Display in *lin* or *log*. Default to *lin*.}

\item{size}{Size of the dots. Default to *0.5*.}

\item{simulated}{Use simulated Blq data? Default to *FALSE*.}

\item{corrected}{Dose the predictions are corrected? Default to *FALSE*.}
}
\description{
This function allows you to have almost a publication ready VPC plot of a monolix project.
}
\examples{
# getwd()
}
\keyword{monolix}
