% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlxpop.xlsx.R
\name{mlxpop.xlsx}
\alias{mlxpop.xlsx}
\title{A function to export Populationparameters.txt file of a Monolix project to xlsx file.}
\usage{
mlxpop.xlsx(
  project.dir = "../monolix/",
  project.name = "",
  estim = "sa",
  caption = "Pop params",
  greek = FALSE
)
}
\arguments{
\item{project.dir}{Absolute or relative name of the folder of the monolix project.}

\item{project.name}{Name of the monolix project file without *.mlxtran* extension.. The project must be in *project.dir*}

\item{estim}{"sa" or "lin". default to "sa".}

\item{caption}{Caption of the table.}

\item{greek}{Convert to greek letters, default to FALSE.}
}
\description{
This function allows you to have almost a publication ready table of the population parameter of a model project.
}
\examples{
# getwd()
}
\keyword{monolix}
